requirements.txt
python-telegram-bot==22.1
python-dotenv==1.0.1
ollama==0.3.3
requests==2.31.0
aiohttp==3.10.10
psutil==5.9.8
feedparser==6.0.11
beautifulsoup4==4.12.3
lxml==5.3.0
pytz==2024.1 